run grid_map.py
You should see a green field filled with tiles
and a 200px field on the right for game and tile info to be added later.

Click on a tile and it will change color, indicating it was selected.
Click on a different tile and the new one will become selected and the
previous one will return to grass color.

Key to the grid_map class is the matrix, a 3-D array for all tiles.
Current info it contains is commented in the init area.
I left most of the testing comments printable. I also left notes
of things yet to do or to be connected once we have object and unit classes.

(It will need to access the new images_lib.py which holds color info
and params.py which holds program specs)

Ignore battle_grid.py